cloud-run-demo-deep-dive-youtube/demo-2-gcp-cloud-storage/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-api-core==1.26.0
google-auth==1.26.1
google-cloud-core==1.6.0
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
